huge fix in foundation service and some minor aditions
the locking was not done right when adding new object
it could happen an deleted memory read while accessing some parts
of the deque holding the objects in some threads while doing a 
push_back in another. now we fully lock the deque.
the only performance degradation is on adding a new object
that must be push_back